COUR-965: Implementing prototype
