+++ to-do-backend/src/main/kotlin/com/yasuo/scopes/prototype/Robot.kt
@@ -1,0 +1,7 @@
+package com.yasuo.scopes.prototype
+
+import io.micronaut.context.annotation.Prototype
+import java.util.*
+
+@Prototype
+data class Robot(val serialNumber:String = UUID.randomUUID().toString())
